+++ jrnl/Journal.py
@@ -11,6 +11,7 @@ try: import parsedatetime.parsedatetime_consts as pdt
 except ImportError: import parsedatetime.parsedatetime as pdt
 import re
 from datetime import datetime
+import dateutil
 import time
 import sys
 try:
@@ -50,9 +51,11 @@ class Journal(object):
         self.search_tags = None  # Store tags we're highlighting
         self.name = name
 
-        journal_txt = self.open()
-        self.entries = self.parse(journal_txt)
-        self.sort()
+        self.open()
+
+    def __len__(self):
+        """Returns the number of entries"""
+        return len(self.entries)
 
     def _colorize(self, string):
         if colorama:
@@ -115,9 +118,10 @@ class Journal(object):
         else:
             with codecs.open(filename, "r", "utf-8") as f:
                 journal = f.read()
-        return journal
+        self.entries = self._parse(journal)
+        self.sort()
 
-    def parse(self, journal):
+    def _parse(self, journal_txt):
         """Parses a journal that's stored in a string and returns a list of entries"""
 
         # Entries start with a line that looks like 'date title' - let's figure out how
@@ -128,7 +132,7 @@ class Journal(object):
         entries = []
         current_entry = None
 
-        for line in journal.splitlines():
+        for line in journal_txt.splitlines():
             try:
                 # try to parse line as date => new entry begins
                 line = line.strip()
@@ -184,7 +188,7 @@ class Journal(object):
     def write(self, filename=None):
         """Dumps the journal into the config file, overwriting it"""
         filename = filename or self.config['journal']
-        journal = "\n".join([e.__unicode__() for e in self.entries])
+        journal = u"\n".join([e.__unicode__() for e in self.entries])
         if self.config['encrypt']:
             journal = self._encrypt(journal)
             with open(filename, 'wb') as journal_file:
@@ -249,7 +253,12 @@ class Journal(object):
         elif isinstance(date_str, datetime):
             return date_str
 
-        date, flag = self.dateparse.parse(date_str)
+        try:
+            date = dateutil.parser.parse(date_str)
+            flag = 1 if date.hour == 0 and date.minute == 0 else 2
+            date = date.timetuple()
+        except:
+            date, flag = self.dateparse.parse(date_str)
 
         if not flag:  # Oops, unparsable.
             try:  # Try and parse this as a single year
@@ -281,20 +290,15 @@ class Journal(object):
         raw = raw.replace('\\n ', '\n').replace('\\n', '\n')
         starred = False
         # Split raw text into title and body
-        title_end = len(raw)
-        for separator in ["\n", ". ", "? ", "! "]:
-            sep_pos = raw.find(separator)
-            if 1 < sep_pos < title_end:
-                title_end = sep_pos
-        title = raw[:title_end+1]
-        body = raw[title_end+1:].strip()
+        sep = re.search("[\n!?.]+", raw)
+        title, body = (raw[:sep.end()], raw[sep.end():]) if sep else (raw, "")
         starred = False
         if not date:
-            if title.find(":") > 0:
-                starred = "*" in title[:title.find(":")]
-                date = self.parse_date(title[:title.find(":")])
+            if title.find(": ") > 0:
+                starred = "*" in title[:title.find(": ")]
+                date = self.parse_date(title[:title.find(": ")])
                 if date or starred:  # Parsed successfully, strip that from the raw text
-                    title = title[title.find(":")+1:].strip()
+                    title = title[title.find(": ")+1:].strip()
             elif title.strip().startswith("*"):
                 starred = True
                 title = title[1:].strip()
@@ -304,25 +308,36 @@ class Journal(object):
         if not date:  # Still nothing? Meh, just live in the moment.
             date = self.parse_date("now")
         entry = Entry.Entry(self, date, title, body, starred=starred)
+        entry.modified = True
         self.entries.append(entry)
         if sort:
             self.sort()
         return entry
 
+    def editable_str(self):
+        """Turns the journal into a string of entries that can be edited
+        manually and later be parsed with eslf.parse_editable_str."""
+        return u"\n".join([e.__unicode__() for e in self.entries])
+
+    def parse_editable_str(self, edited):
+        """Parses the output of self.editable_str and updates it's entries."""
+        mod_entries = self._parse(edited)
+        # Match those entries that can be found in self.entries and set
+        # these to modified, so we can get a count of how many entries got
+        # modified and how many got deleted later.
+        for entry in mod_entries:
+            entry.modified = not any(entry == old_entry for old_entry in self.entries)
+        self.entries = mod_entries
 
 class DayOne(Journal):
     """A special Journal handling DayOne files"""
     def __init__(self, **kwargs):
         self.entries = []
+        self._deleted_entries = []
         super(DayOne, self).__init__(**kwargs)
 
     def open(self):
-        files = [os.path.join(self.config['journal'], "entries", f) for f in os.listdir(os.path.join(self.config['journal'], "entries"))]
-        return files
-
-    def parse(self, filenames):
-        """Instead of parsing a string into an entry, this method will take a list
-        of filenames, interpret each as a plist file and create a new entry from that."""
+        filenames = [os.path.join(self.config['journal'], "entries", f) for f in os.listdir(os.path.join(self.config['journal'], "entries"))]
         self.entries = []
         for filename in filenames:
             with open(filename, 'rb') as plist_entry:
@@ -333,34 +348,97 @@ class DayOne(Journal):
                     timezone = pytz.timezone(util.get_local_timezone())
                 date = dict_entry['Creation Date']
                 date = date + timezone.utcoffset(date)
-                entry = self.new_entry(raw=dict_entry['Entry Text'], date=date, sort=False)
-                entry.starred = dict_entry["Starred"]
+                raw = dict_entry['Entry Text']
+                sep = re.search("[\n!?.]+", raw)
+                title, body = (raw[:sep.end()], raw[sep.end():]) if sep else (raw, "")
+                entry = Entry.Entry(self, date, title, body, starred=dict_entry["Starred"])
                 entry.uuid = dict_entry["UUID"]
                 entry.tags = dict_entry.get("Tags", [])
-        # We're using new_entry to create the Entry object, which adds the entry
-        # to self.entries already. However, in the original Journal.__init__, this
-        # method is expected to return a list of newly created entries, which is why
-        # we're returning the obvious.
-        return self.entries
+                self.entries.append(entry)
+        self.sort()
 
     def write(self):
         """Writes only the entries that have been modified into plist files."""
         for entry in self.entries:
-            # Assumption: since jrnl can not manipulate existing entries, all entries
-            # that have a uuid will be old ones, and only the one that doesn't will
-            # have a new one!
-            if not hasattr(entry, "uuid"):
+            if entry.modified:
+                if not hasattr(entry, "uuid"):
+                    entry.uuid = uuid.uuid1().hex
                 utc_time = datetime.utcfromtimestamp(time.mktime(entry.date.timetuple()))
-                new_uuid = uuid.uuid1().hex
-                filename = os.path.join(self.config['journal'], "entries", new_uuid+".doentry")
+                filename = os.path.join(self.config['journal'], "entries", entry.uuid+".doentry")
                 entry_plist = {
                     'Creation Date': utc_time,
                     'Starred': entry.starred if hasattr(entry, 'starred') else False,
                     'Entry Text': entry.title+"\n"+entry.body,
                     'Time Zone': util.get_local_timezone(),
-                    'UUID': new_uuid,
+                    'UUID': entry.uuid,
                     'Tags': [tag.strip(self.config['tagsymbols']) for tag in entry.tags]
                 }
-                # print entry_plist
-
                 plistlib.writePlist(entry_plist, filename)
+        for entry in self._deleted_entries:
+            filename = os.path.join(self.config['journal'], "entries", entry.uuid+".doentry")
+            os.remove(filename)
+
+    def editable_str(self):
+        """Turns the journal into a string of entries that can be edited
+        manually and later be parsed with eslf.parse_editable_str."""
+        return u"\n".join(["# {0}\n{1}".format(e.uuid, e.__unicode__()) for e in self.entries])
+
+    def parse_editable_str(self, edited):
+        """Parses the output of self.editable_str and updates it's entries."""
+        # Method: create a new list of entries from the edited text, then match
+        # UUIDs of the new entries against self.entries, updating the entries
+        # if the edited entries differ, and deleting entries from self.entries
+        # if they don't show up in the edited entries anymore.
+        date_length = len(datetime.today().strftime(self.config['timeformat']))
+
+        # Initialise our current entry
+        entries = []
+        current_entry = None
+
+        for line in edited.splitlines():
+            # try to parse line as UUID => new entry begins
+            line = line.strip()
+            m = re.match("# *([a-f0-9]+) *$", line.lower())
+            if m:
+                if current_entry:
+                    entries.append(current_entry)
+                current_entry = Entry.Entry(self)
+                current_entry.modified = False
+                current_entry.uuid = m.group(1).lower()
+            else:
+                try:
+                    new_date = datetime.strptime(line[:date_length], self.config['timeformat'])
+                    if line.endswith("*"):
+                        current_entry.starred = True
+                        line = line[:-1]
+                    current_entry.title = line[date_length+1:]
+                    current_entry.date = new_date
+                except ValueError:
+                    if current_entry:
+                        current_entry.body += line + "\n"
+
+        # Append last entry
+        if current_entry:
+            entries.append(current_entry)
+
+        # Now, update our current entries if they changed
+        for entry in entries:
+            entry.parse_tags()
+            matched_entries = [e for e in self.entries if e.uuid.lower() == entry.uuid]
+            if matched_entries:
+                # This entry is an existing entry
+                match = matched_entries[0]
+                if match != entry:
+                    self.entries.remove(match)
+                    entry.modified = True
+                    self.entries.append(entry)
+            else:
+                # This entry seems to be new... save it.
+                entry.modified = True
+                self.entries.append(entry)
+        # Remove deleted entries
+        edited_uuids = [e.uuid for e in entries]
+        self._deleted_entries = [e for e in self.entries if e.uuid not in edited_uuids]
+        self.entries[:] = [e for e in self.entries if e.uuid in edited_uuids]
+        return entries
+